+++ spec/generators/install_generator_spec.rb
@@ -15,14 +15,30 @@ RSpec.describe PaperTrail::InstallGenerator, type: :generator do
     end
 
     it "generates a migration for creating the 'versions' table" do
+      expected_parent_class = lambda {
+        old_school = "ActiveRecord::Migration"
+        ar_version = ActiveRecord::VERSION
+        if ar_version::MAJOR >= 5
+          format("%s[%d.%d]", old_school, ar_version::MAJOR, ar_version::MINOR)
+        else
+          old_school
+        end
+      }.call
+      expected_create_table_options = lambda {
+        if described_class::MYSQL_ADAPTERS.include?(ActiveRecord::Base.connection.class.name)
+          ', { options: "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci" }'
+        else
+          ""
+        end
+      }.call
       expect(destination_root).to(
         have_structure {
           directory("db") {
             directory("migrate") {
               migration("create_versions") {
-                contains "class CreateVersions"
+                contains("class CreateVersions < " + expected_parent_class)
                 contains "def change"
-                contains "create_table :versions"
+                contains "create_table :versions#{expected_create_table_options}"
               }
             }
           }